Fix the TV streams
This preserves the alleged new format of the stream that sends "fragments",
and also preserves the AES-128 encryption used which VLC (and possibly other)
players don't have problems with playing.

This commit also fixes some issues with the code that will make the
project easier to configure i.e. JSON data is not serialized twice, etc.
and the project is using uv, so you can do `uv run main.py`.